requirements.txt
attr==0.3.2
bpython==0.24
channels==4.0.0
ConfigParser==6.0.1
contextlib2==21.6.0
coreschema==0.0.4
cryptography>=3.2.1
cx_Oracle==8.3.0
dj_rest_auth==5.1.0
django_allauth==0.61.1
docutils==0.20.1
filelock==3.13.3
HTMLParser==0.0.2
importlib_metadata==4.6.4
inflection==0.5.1
ipdb==0.13.13
ipython==8.12.3
ipywidgets==8.1.2
keyring==23.5.0
Markdown==3.6
mock==5.1.0
numpy==1.26.4
oracledb==2.1.1
Pillow==10.2.0
protobuf==5.26.1
psycopg==3.1.18
psycopg2_binary==2.9.9
pylibmc==1.6.3
pymemcache==4.0.0
pyOpenSSL==24.1.0
pytz==2024.1
pywatchman==2.0.0
PyYAML==6.0.1
railroad==0.5.0
Sphinx==7.2.6
